Web Service Endpoint/requirements.txt
numpy
pandas
scikit-learn
scipy
joblib
category-encoders
opencv-python
psycopg2-binary
sqlalchemy
boto3
urllib3>=1.24.2
flask>=0.12.3
flask-cors
flask-sqlalchemy
python-decouple 
python-dotenv 
gunicorn